+++ packages/application/test/goal-cancellation-service.unit.test.ts
@@ -35,6 +35,7 @@ describe('GoalCancellationService', () => {
         listActiveByGoalId: () =>
           Promise.resolve([instance('instance-1', 'plan-1'), instance('instance-2', 'plan-2')]),
         countNodeEvents: () => Promise.resolve(0),
+        listNodeEvents: () => Promise.resolve([]),
         saveInstance: () => Promise.resolve(),
         saveNodeEvents: () => Promise.resolve(),
       },